+++ version3/java/ROM_NUMS256W_64.java
@@ -33,7 +33,7 @@ public class ROM
 // Curve
 
 	public static final int CURVE_Cof_I= 1;
-	public static final int[] CURVE_Cof= {0x1L,0x0L,0x0L,0x0L,0x0L};
+	public static final long[] CURVE_Cof= {0x1L,0x0L,0x0L,0x0L,0x0L};
 	public static final int CURVE_A= -3;
 	public static final int CURVE_B_I= 152961;
 	public static final long[] CURVE_B= {0x25581L,0x0L,0x0L,0x0L,0x0L};
